ui updated.

+++ iColor/detailMixView.m
@@ -99,11 +99,11 @@
     saveButton.backgroundColor = buttonColor.myUIColor;
     
     //labels setting.
-    [redLabel setText:[NSString stringWithFormat:  @"     Red:              %li", thisColor.rValue]];
+    [redLabel setText:[NSString stringWithFormat:  @"     Red:              %li", (long)thisColor.rValue]];
     [redLabel setBackgroundColor:[UIColor colorWithRed:231/255.0 green:76/255.0 blue:60/255.0 alpha:1]];
-    [greenLabel setText:[NSString stringWithFormat:@"     Green:          %li", thisColor.gValue]];
+    [greenLabel setText:[NSString stringWithFormat:@"     Green:          %li", (long)thisColor.gValue]];
     [greenLabel setBackgroundColor:[UIColor colorWithRed:46/255.0 green:204/255.0 blue:113/255.0 alpha:1]];
-    [blueLabel setText:[NSString stringWithFormat: @"     Blue:             %li", thisColor.bValue]];
+    [blueLabel setText:[NSString stringWithFormat: @"     Blue:             %li", (long)thisColor.bValue]];
     [blueLabel setBackgroundColor:[UIColor colorWithRed:52/255.0 green:152/255.0 blue:219/255.0 alpha:1]];
     [alphaLabel setText:[NSString stringWithFormat:@"     Alpha:          %f", thisColor.brightness]];
     [alphaLabel setBackgroundColor:[UIColor colorWithRed:155/255.0 green:89/255.0 blue:182/255.0 alpha:1]];
